+++ tests/test_helpers.py
@@ -38,3 +38,33 @@ def test_median_3():
     values = [1, 2, 3]
     result = helpers.median(values)
     assert result == 2
+
+
+def test_sanitize_path_blank():
+    path = ""
+    result = helpers.sanitize_path(path)
+    assert result == ""
+
+
+def test_sanitize_path_0():
+    path = "test"
+    result = helpers.sanitize_path(path)
+    assert result == "test"
+
+
+def test_sanitize_path_1():
+    path = "te%st"
+    result = helpers.sanitize_path(path)
+    assert result == "te_st"
+
+
+def test_sanitize_path_hyphen():
+    path = "te-st"
+    result = helpers.sanitize_path(path)
+    assert result == "te-st"
+
+
+def test_sanitize_path_many():
+    path = "!@#$%^&*()=+`~<>?/"';:'
+    result = helpers.sanitize_path(path)
+    assert result == "_" * len(path)